refactor: extract GET `/api/data` to file
